CineCompassBackend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
pytest==8.3.4
httpx==0.28.0
python-multipart==0.0.19
pydantic==2.10.2
numpy==1.26.2
requests==2.32.3
sqlalchemy==2.0.36